requirements.txt
streamlit
PyPDF2
spacy
python-docx
fpdf
openai

